requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-config-file==0.6.0
click-plugins==1.1.1
cloudpathlib==0.16.0
confection==0.1.4
configobj==5.0.8
cymem==2.0.8
decorator==5.1.1
distro==1.9.0
Django==5.0.3
djangorestframework==3.15.1
emoji==2.11.0
emot==3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
genai==2.1.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.6
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==0.27.10
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.1
rsa==4.9
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
tabulate==0.9.0
textblob==0.18.0.post0
thinc==8.2.3
tiktoken==0.3.3
tqdm==4.66.2
traitlets==5.14.2
twarc==2.14.0
tweepy==4.14.0
twitter==1.19.6
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
whitenoise==6.6.0
xtract==0.1a3
yarl==1.9.4
